backend/requirements.txt
Flask==2.2.2
pymongo==4.2.0
python-dotenv==0.19.2
requests==2.28.1
beautifulsoup4==4.11.1
pandas==1.5.3  